requirements.txt
numpy>=1.19.0
scipy>=1.5.0
matplotlib>=3.3.0
scikit-image>=0.17.0
scikit-learn>=0.23.0
open3d>=0.12.0
vtk>=9.0.0
pyvista>=0.30.0
SimpleITK>=2.0.0
pydicom>=2.0.0
nibabel>=3.2.0
pynrrd>=0.4.0
pillow>=8.0.0
numpy-stl>=2.16.0 